Adding the new Button props
